Fix image controller deleteImage function

+++ src/main/java/ImageHoster/controller/ImageController.java
@@ -153,9 +153,21 @@ public class ImageController {
     //The method calls the deleteImage() method in the business logic passing the id of the image to be deleted
     //Looks for a controller method with request mapping of type '/images'
     @RequestMapping(value = "/deleteImage", method = RequestMethod.DELETE)
-    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId) {
-        imageService.deleteImage(imageId);
-        return "redirect:/images";
+    public String deleteImageSubmit(@RequestParam(name = "imageId") Integer imageId, HttpSession session, RedirectAttributes redirect) {
+        Image image = imageService.getImage(imageId);
+        User imageCreator = image.getUser();
+        User loggedInUser = (User) session.getAttribute("loggeduser");
+
+        if(!imageCreator.getUsername().equalsIgnoreCase(loggedInUser.getUsername())){
+            String error = "Only the owner of the image can delete the image";
+            String imgTit = image.getTitle();
+            redirect.addAttribute("deleteError", error).addFlashAttribute("deleteError", error);
+            return "redirect:images/"+imageId+"/"+imgTit;
+        }
+        else {
+            imageService.deleteImage(imageId);
+            return "redirect:/images";
+        }
     }
